api/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.7.1
sqlparse==0.4.3
psycopg2-binary==2.9.6
modbus-tk==1.1.3
loguru==0.7.0

python-snap7~=1.3
cprint~=1.2.2